Add a check for objective-c class names when checking the reference counts in script tasks to avoid warnings when find is used to retrive Orca objects.
Add a "read only" mode to the script tasks.  When the folder icon is opened from the script task, "load as read only" is now an option.  This allows the script to be loaded from a file, and, once loaded, the GUI fields which would allow the script body to be edited are disabled.  Once a script is read only, it cannot be reverted from the GUI.  It can still be set to editable using objective-c methods from another script task.