Enable energy histograms for the SIS3316 in the data monitor.
This requires that either the MAW values or the start/end energy options are enabled in the dialog for the card.  By default, the MAW values are used, but, if they are not in the data stream, the start/end energy values are uesd.  If neither of these options in the dialog are checked, no histograms produced.

The change to the OR1DHistoPlot is required to extend the x-axis scale to the range required for plotting the SIS3316 spectrum.