Initial implementation of the LEGEND-200 experiment object.
The L200 object currently supports channel maps for the various subsystems and displays of the rates in crate and experiment views.

Add the ORAdcInfoProviding protocol to the FlashCam ADC card for use in the L200 object.  Modify the FlashCam ADC card to keep track of both the number of waveforms recorded for each channel and the number of waveforms that crossed threshold on each channel.  Both rates are displayed in the Rates tab for the object.

Add a secondary data source to ORValueBar and ORValueBarGroupView to allow a pair of bars to be displayed in the same value bar for each bar in the group.  This is used in the FlashCam ADC card to display the rate of triggers and the total rate for each channel.