ORHVcRIOModel: More digits
Add digits to values in `ORHVcRIOModel`. At KATRIN, we need to set some values with further precision than is possible with the HV object at the moment - 8 digits should be enough.
This is relevant especially because of `SOLL_EHV#postReg`, in which this discretization limited our ability to set this value precisely.
If there is no reason against using the maximal amount of digits, we could also go to `"%lf` as is done in `setMeasuredValue`?