.Always write alarms to log
Previously, alarms were only written to log if they were not already posted. This is confusing since this means some alarms are sent via mail but not logged if operators decide to not acknowledge an old alarm. Therefore, this commit changes the alarms so they always occur in the status log.